[refactor] ProductRepository 의존성 분리
