+++ src/main/java/kr/hhplus/be/server/product/application/ProductRepository.java
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.product.application;
+
+import kr.hhplus.be.server.product.infrastructure.entity.ProductEntity;
+
+public interface ProductRepository{
+    ProductEntity findById(Long productId);
+    ProductEntity save(ProductEntity productEntity);
+}
