+++ src/main/java/kr/hhplus/be/server/product/application/ProductService.java
@@ -1,22 +1,19 @@
 package kr.hhplus.be.server.product.application;
 
-import kr.hhplus.be.server.exception.custom.ResourceNotFoundException;
 import kr.hhplus.be.server.product.application.dto.ProductServiceDTO.ProductResult;
-import kr.hhplus.be.server.product.infrastructure.repository.ProductRepository;
+import kr.hhplus.be.server.product.infrastructure.entity.ProductEntity;
 import kr.hhplus.be.server.product.mapper.ProductMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-@RequiredArgsConstructor
 @Service
+@RequiredArgsConstructor
 public class ProductService {
 
-    private final ProductRepository productRepository;
+    private ProductRepository productRepository;
 
     public ProductResult findProduct(Long productId) {
-        return productRepository.findById(productId)
-                .map(ProductMapper.INSTANCE::entityToDomain)
-                .map(ProductResult::from)
-                .orElseThrow(ResourceNotFoundException::new);
+        ProductEntity productEntity = productRepository.findById(productId);
+        return ProductResult.from(ProductMapper.INSTANCE.entityToDomain(productEntity));
     }
 }
