--- src/main/java/kr/hhplus/be/server/product/infrastructure/repository/ProductRepository.java
@@ -1,10 +1,0 @@
-package kr.hhplus.be.server.product.infrastructure.repository;
-
-import kr.hhplus.be.server.product.infrastructure.entity.ProductEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
-
-}
