+++ src/test/java/kr/hhplus/be/server/product/presentation/ProductControllerTest.java
@@ -1,14 +1,10 @@
 package kr.hhplus.be.server.product.presentation;
 
-import kr.hhplus.be.server.product.application.ProductService;
 import kr.hhplus.be.server.product.common.ProductState;
 import kr.hhplus.be.server.product.infrastructure.entity.ProductEntity;
-import kr.hhplus.be.server.product.infrastructure.repository.ProductRepository;
-import kr.hhplus.be.server.product.presentation.ApiResponse;
-import org.junit.jupiter.api.BeforeEach;
+import kr.hhplus.be.server.product.application.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
